+++ tools/OpenSplat/opensplat.hpp
@@ -1,6 +1,0 @@
-#ifndef OPENSPLAT_H
-#define OPENSPLAT_H
-
-#include "model.hpp"
-
-#endif