+++ tools/OpenSplat/tile_bounds.hpp
@@ -1,8 +1,0 @@
-#ifndef TILE_BOUNDS_H
-#define TILE_BOUNDS_H
-
-#include <tuple> 
-
-typedef std::tuple<int, int, int> TileBounds; 
-
-#endif